convert indexes to integers to prepare for switch to Uint16Array
We'll do a breaking change to music-routes-data if this all works out so
we can then get rid of the conversions.

Maybe add a check to the music-routes-data tests to make sure indexes
are within the range of Uint16Array so we can do another breaking change
if we ever need to move to Uint32Array.